+++ packages/components/src/Message/Message.stories.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import type { Meta, ComponentStory } from '@storybook/react';
 import { Message } from './Message';
 import { Text } from '../Text';
+import { Box } from '@marigold/system';
 
 export default {
   title: 'Components/Message',
@@ -26,4 +27,15 @@ export const Basic: ComponentStory<typeof Message> = args => (
   <Message {...args}>
     <Text>Hello, I am a simple message.</Text>
   </Message>
+);
+
+export const MultiLineTitle: ComponentStory<typeof Message> = () => (
+  <Box css={{ width: 250 }}>
+    <Message
+      messageTitle="Hey! You! I am an info box! Please notice me, it might help you!"
+      variant="info"
+    >
+      <Text>I am really not that good at righting copy texts, sorry.</Text>
+    </Message>
+  </Box>
 );