+++ themes/theme-b2b/src/components/Text.style.ts
@@ -1,7 +1,3 @@
 import type { Theme } from '@marigold/components';
 
-export const Text: Theme['components']['Text'] = {
-  base: {
-    fontSize: 'xxsmall',
-  },
-};
+export const Text: Theme['components']['Text'] = {};