+++ themes/theme-unicorn/src/components/ListBox.style.ts
@@ -10,6 +10,9 @@ export const ListBox: Theme['components']['ListBox'] = {
       borderColor: 'gray40',
       borderRadius: 'small',
 
+      height: '75vh',
+      overflow: 'hidden auto',
+
       '&:error': {
         borderColor: 'error',
       },