+++ config/storybook/CHANGELOG.md
@@ -1,5 +1,57 @@
 # @marigold/storybook-config
 
+## 0.7.20
+
+### Patch Changes
+
+- Updated dependencies [[`1399a66b`](https://github.com/marigold-ui/marigold/commit/1399a66ba8cacb00fca6209d380a7b739ed94e7d)]:
+  - @marigold/theme-b2b@16.0.4
+  - @marigold/theme-core@16.0.4
+  - @marigold/theme-unicorn@13.0.4
+  - @marigold/components@4.1.4
+
+## 0.7.19
+
+### Patch Changes
+
+- Updated dependencies [[`9f12ffec`](https://github.com/marigold-ui/marigold/commit/9f12ffecb4e7fdf0bcc407a9a7867ff4c1263a99)]:
+  - @marigold/theme-b2b@16.0.3
+  - @marigold/theme-core@16.0.3
+  - @marigold/theme-unicorn@13.0.3
+  - @marigold/components@4.1.3
+
+## 0.7.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @marigold/components@4.1.2
+  - @marigold/theme-b2b@16.0.2
+  - @marigold/theme-core@16.0.2
+  - @marigold/theme-unicorn@13.0.2
+
+## 0.7.17
+
+### Patch Changes
+
+- Updated dependencies [[`0208fb01`](https://github.com/marigold-ui/marigold/commit/0208fb0111f44ed2de78098c1cf4757a55302048), [`9673ba62`](https://github.com/marigold-ui/marigold/commit/9673ba622011cb3cc748dba14f3c533160c95eee)]:
+  - @marigold/components@4.1.1
+  - @marigold/theme-b2b@16.0.1
+  - @marigold/theme-core@16.0.1
+  - @marigold/theme-unicorn@13.0.1
+
+## 0.7.16
+
+### Patch Changes
+
+- [#2627](https://github.com/marigold-ui/marigold/pull/2627) [`d250fc00`](https://github.com/marigold-ui/marigold/commit/d250fc0041f2beca498107d3e60e2d50e9ffb293) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update babel monorepo
+
+- Updated dependencies [[`40aeefd7`](https://github.com/marigold-ui/marigold/commit/40aeefd704e4ac23b36cb83ade928b19bdaf68c0), [`2e98753d`](https://github.com/marigold-ui/marigold/commit/2e98753d80af8a00a2b8c9adb0d3e08ee631ea4b), [`f764d3a0`](https://github.com/marigold-ui/marigold/commit/f764d3a080f075fe4e9b0cea30ded53a460689b4), [`473ae72b`](https://github.com/marigold-ui/marigold/commit/473ae72bc5ea11c8092a613b76500140f78cd12f), [`7a9129c1`](https://github.com/marigold-ui/marigold/commit/7a9129c1c2d2ada75fa5cd793874ab0b7db8d48b), [`727460fc`](https://github.com/marigold-ui/marigold/commit/727460fcd2dce035b18a539512ecf89f1ab7a11b)]:
+  - @marigold/components@4.1.0
+  - @marigold/theme-b2b@16.0.0
+  - @marigold/theme-core@16.0.0
+  - @marigold/theme-unicorn@13.0.0
+
 ## 0.7.15
 
 ### Patch Changes