+++ themes/theme-unicorn/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @marigold/theme-unicorn
 
+## 13.0.4
+
+### Patch Changes
+
+- [#2666](https://github.com/marigold-ui/marigold/pull/2666) [`1399a66b`](https://github.com/marigold-ui/marigold/commit/1399a66ba8cacb00fca6209d380a7b739ed94e7d) Thanks [@sarahgm](https://github.com/sarahgm)! - chore: change menu/select height to 75vh
+
+- Updated dependencies []:
+  - @marigold/components@4.1.4
+
+## 13.0.3
+
+### Patch Changes
+
+- [#2663](https://github.com/marigold-ui/marigold/pull/2663) [`9f12ffec`](https://github.com/marigold-ui/marigold/commit/9f12ffecb4e7fdf0bcc407a9a7867ff4c1263a99) Thanks [@benediktgrether](https://github.com/benediktgrether)! - fix: set height to 25vh and overflow to auto for select box and menu
+
+- Updated dependencies []:
+  - @marigold/components@4.1.3
+
+## 13.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @marigold/components@4.1.2
+
+## 13.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`0208fb01`](https://github.com/marigold-ui/marigold/commit/0208fb0111f44ed2de78098c1cf4757a55302048)]:
+  - @marigold/components@4.1.1
+
+## 13.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`40aeefd7`](https://github.com/marigold-ui/marigold/commit/40aeefd704e4ac23b36cb83ade928b19bdaf68c0), [`2e98753d`](https://github.com/marigold-ui/marigold/commit/2e98753d80af8a00a2b8c9adb0d3e08ee631ea4b), [`f764d3a0`](https://github.com/marigold-ui/marigold/commit/f764d3a080f075fe4e9b0cea30ded53a460689b4), [`473ae72b`](https://github.com/marigold-ui/marigold/commit/473ae72bc5ea11c8092a613b76500140f78cd12f), [`7a9129c1`](https://github.com/marigold-ui/marigold/commit/7a9129c1c2d2ada75fa5cd793874ab0b7db8d48b), [`727460fc`](https://github.com/marigold-ui/marigold/commit/727460fcd2dce035b18a539512ecf89f1ab7a11b)]:
+  - @marigold/components@4.1.0
+
 ## 12.0.0
 
 ### Patch Changes